Renaming existing k8s nodeattestor to k8s/sat
