+++ pkg/agent/catalog/catalog.go
@@ -12,7 +12,7 @@ import (
 	"github.com/spiffe/spire/pkg/agent/plugin/nodeattestor/azure"
 	"github.com/spiffe/spire/pkg/agent/plugin/nodeattestor/gcp"
 	"github.com/spiffe/spire/pkg/agent/plugin/nodeattestor/jointoken"
-	k8s_na "github.com/spiffe/spire/pkg/agent/plugin/nodeattestor/k8s"
+	k8s_na_sat "github.com/spiffe/spire/pkg/agent/plugin/nodeattestor/k8s/sat"
 	"github.com/spiffe/spire/pkg/agent/plugin/nodeattestor/x509pop"
 	"github.com/spiffe/spire/pkg/agent/plugin/workloadattestor/docker"
 	k8s_wa "github.com/spiffe/spire/pkg/agent/plugin/workloadattestor/k8s"
@@ -55,7 +55,7 @@ var (
 			"gcp_iit":    nodeattestor.NewBuiltIn(gcp.NewIITAttestorPlugin()),
 			"x509pop":    nodeattestor.NewBuiltIn(x509pop.New()),
 			"azure_msi":  nodeattestor.NewBuiltIn(azure.NewMSIAttestorPlugin()),
-			"k8s_sat":    nodeattestor.NewBuiltIn(k8s_na.NewSATAttestorPlugin()),
+			"k8s_sat":    nodeattestor.NewBuiltIn(k8s_na_sat.NewSATAttestorPlugin()),
 		},
 		WorkloadAttestorType: {
 			"k8s":    workloadattestor.NewBuiltIn(k8s_wa.New()),
