+++ pkg/server/plugin/nodeattestor/k8s/sat/sat.go
@@ -1,4 +1,4 @@
-package k8s
+package sat
 
 import (
 	"context"
